General: - code cleanup and refactoring, - misc bugfixes.
CPU:
- replaced current addressing mode detection with something much more reliable (especially for unofficial opcodes),
- added more info for debugging logs.

Boards:
- fixed masking for UxROM.java

APU:
- fixed unnecessary DMC DMA cycle stealing,
- started connecting audio to GUI.

GUI:
- now at least a background shows up! yay for progress,
- started working on saving preferences,
- connected joypads to GUI.

+++ src/nesimulare/core/memory/CPUMemory.java
@@ -66,7 +66,7 @@ public class CPUMemory extends Memory  {
         } else if (address < 0x2000) {
             wram[address & 0x7FF] = data;
         } else if (address < 0x4000 || address == 0x4014) {
-            nes.ppu.write(address, data);
+            nes.ppu.write(address, (byte)data);
         } else if (address >= 0x4000 && address <= 0x4018) {
             nes.apu.write(address, data);
         } else if (address < 0x6000) {
